requirements.txt
Django~=4.2.1
mysqlclient~=2.1.1
google-auth~=2.22.0
requests~=2.31.0
canvasapi~=3.2.0
Pillow~=10.0.0
spotipy~=2.23.0
